:fishing_pole_and_fish: Remove unnecessary casting
Addresses https://git.io/vMHw8.

+++ Management.Database/Models/ManaDbVote.cs
@@ -45,9 +45,9 @@ namespace Management.Database.Models
         {
             return new ManaVote
             {
-                Id = (int) Id,
-                OrgId = (int) OrgId,
-                CreatedBy = (int) CreatedBy,
+                Id = Id,
+                OrgId = OrgId,
+                CreatedBy = CreatedBy,
                 Name = Name,
                 CreationDate = CreationDate,
                 ExpiryDate = ExpiryDate,
